♻️ Refactor: 캔버스들 분리
