+++ Frontend/src/pages/StudioPage/Canvas/DrawCanvas.tsx
