+++ Frontend/src/pages/StudioPage/Canvas/ImageTextCanvas.tsx
