+++ Frontend/src/pages/StudioPage/Canvas/StreamCanvas.tsx
